Simple Fix to Error Reporting on Var Declarations
Previously if there was a type validation error found in the rhs expr on var declaration, then the declared var would be placed in the symbol table with a `null` type. This led to NPEs elsewhere so Claro actually wasn't able to gracefully find more errors after that issue. This is now fixed with the addition of explicitly modelling an `UNKNOWABLE` "type" that the var will be declared to have. This way any further usages of the variable will result in a clear message that will indicate it's failing b/c of a prior type validation error. (In fact, in the future, I'll likely filter out those particular messages so the users can focus just on root causes).