Fix Re-entrancy Bug in ContractFunctionCallExpr.java
This was breaking contract function calls made within generic functions because when it was monomorphized the type validation being performed again over the same node with the new concrete types was effectively not being treated as a clean type checking run. This change fixes that so that type checking gets a clean run for each set of concrete types.

Unfortunately, Claro's fundamental design makes it so that each AST node is susceptible to this problem. If/when I do a rewrite in Claro, I must ensure idempotency of type validation so that I can monomorphize fearlessly. For now, I'll just continue ahead with care.