Contracts Now Support "Implied Types"!
This can be taken as an analogue of Rust/Swift "Associated Types" with naming and syntax that I personally believe is more appropriate.

The introduction of the sigil `=>` for use *only* in Contract Type Param declaration lists, any type param(s) listed after the `=>` will be considered "implied" by the type param(s) preceding. This choice of operator was very intentional, and yes, it does completely retain its mathematical meaning of "logical implication".