Language Level Mutability/Immutability Enforcement on Collection Types!
This was actually a shockingly simple change implementation-wise, but this has massive effects language wide. Now, all Claro collection types default to being immutable unless explicitly annotated as being mutable. This is already powerful for the simple fact alone that the language doesn't require libraries to be implemented for things like ImmutableList vs ArrayList in Java, the language comes with this concept built in.

However, the massive power of this is not simply a library concern. Much more important is the fact that Claro will be able to track the possibility of mutation at the language level and provide static enforcement guarantees that state that is shared between threads are provably deeply immutable. Stay tuned!