Require Mutable Collection Literals to be Explicitly Annotated As Such
This is a simple change that again has large perceived change on the language itself. Now, rather than Claro inferring whether the literal `[1]` refers to a mutable or immutable collection based on the type contextually asserted (and when not asserted defaulting to immutable), Claro now just requires users to be explicit in order to get a mutable collection. Now, you'd be forced to explicitly write `mut [1]` to get a mutable list. This avoids making Claro hard to understand when you're reading it, and also allows Claro itself to avoid resorting to internal trickery to handle inferring the type correctly in all cases.