First Step Towards User-Defined Types!
It's hard to believe, but Claro hasn't supported user-defined types all the way up to this point. This has been for good reason though, as when you really stop and think about it *ALL DATA* is at the end of the day just some structured combination of builtin types (and this is true regardless of the language you think of)! However, defining new types is a very powerful way of enforcing *SEMANTIC* distinctions between values that should be *INTERPRETED* differently. For example, `int` and `EvenInt` might very well be *represented in memory* using the same underlying encoding, but their *semantic interpretations* are very different, as `EvenInt` is imposing constraints on the set of `int`s that you could be dealing with. *This* is the motivation for having user-defined types in the language at all - it's to say that you're imposing some new constraints and/or semantic interpretation over another type.

Claro now supports the following syntax:

```
newtype MyInt : int
var x : MyInt = MyInt(1); # Using auto-generated default constructor.
var unwrapped : int = unwrap(x); # Using builtin `unwrap()` function to reinterpret as the wrapped type.
```

In the following changes, there will be some significant enhancements. First will be the ability to constrain the usage of the auto-generated constructor, and the unwrap() function to a single fixed set of procedures. This may sound odd, but it will be a very powerful constraint that will enable the modeling of types which cannot possibly represent values that violate their semantic constraints.