User-Defined Type `initializers` Blocks!
Claro now allows you to constrain the usage of the auto-generated constructor function to a single fixed set of procedures known as "initializers" by defining them within an initializers block for the custom type. This may sound odd, but it is a very powerful constraint that will enable the modeling of types which cannot possibly represent values that violate their semantic constraints.

For example, the below code defines a custom type `EvenInt` that is *represented* as an int at a storage level, but it has the added compile-time validated constraint that you can only retrieve an instance of this type via the initializer function `GetValidatedEvenInt()` which in turn enforces the semantic constrain that it is impossible to represent any `EvenInt` wrapping an odd `int`!

```
newtype NothingType : boolean
initializers NothingType {
  provider Nothing() -> NothingType {
    return NothingType(false);
  }
}

newtype EvenInt : int

initializers EvenInt {
  function GetValidatedEvenInt(x: int) -> oneof<EvenInt, NothingType> {
    if (x % 2 == 1) {
      return Nothing();
    }
    return EvenInt(x);
  }
}
```