User-Defined Type `unwrappers` Blocks!
Claro now allows you to constrain the usage of the auto-generated `unwrap()` function to a single fixed set of procedures known as "unwrappers" by defining them within an unwrappers block for the custom type. This may sound odd, but it is a powerful constraint that will enable the modeling of types which model non-trivial semantic encodings, but are guaranteed to decode themselves properly when converting back to simpler representations.

Check out the complex CSV -> FasterLookupEncodedString example in `user_defined_types.claro`!