Parameterized User-Defined Types!
Claro now supports the definition of custom parameterized types! This is the final major step in support for custom types!!

It's now possible to define types like the following:

```
newtype Node<T> : tuple<T, oneof<T, NothingType>, oneof<T, NothingType>>
newtype List<T> : [T]
newtype Maybe<T> : oneof<T, NothingType>
```

One of the coolest things about this entire implementation is that the auto-generated default constructor is actually defined as a single generic function, and then with minimal changes type checking is completely handled by Claro's existing handling of generic functions. Neat!

Aside: One massive annoyance came up during this which was stumbling across the fact that the grammar would become ambiguous with the newly supported syntax for parameterized custom types. The issue was (in my mind) a conflict between a few separate grammar rules: parenthesized_expr, cast_expr, and builtin_type (but the issue really revealed itself with less_than_bool_expr).

The issue was that with only a small lookahead configured for the JCUP parser generator, parsing was confusing the following boolean expr as if it were a cast, and was expecting a builtin_type when instead it was receiving LEN:

`while ((i < len(l)) and ...)`

The key here was that, for all JCUP knew, the substring `(i < len` could be the beginning of a cast for a custom parameterized type `i<T>` being used in a cast of the form `(i<SomeType>) someExpr`.

So, to solve this, I've actually decided to scrap the original cast syntax in favor of a functional form `cast(castedType, someExpr)` so now there's no way for this to be ambiguous with any other boolean expr. I've made the corresponding refactorings in all affected .claro files in this same CL because this change was caused by the introduction of parameterized types.