Structs!
It's pretty crazy that it's taken this long to *do structs right* but here we finally are! I've thrown away the previous hacked together approach from around 1.5 years ago which didn't at all fit with the language that Claro has become. Now, Claro Structs are an incredibly ergonomic tool that enable some very expressive code w/o requiring the overhead of other languages (like in Java which would require a new class definition which is decidedly not a straightforward process and is usually avoided). In Claro, I believe you'll see heavy uses of Structs as they are as ergonomic as Tuples, but with named fields.