Update to Automatic Error Propagation
Previously, in the first pass at implementing error propagation in Claro, I implemented it as an Expr that would essentially wrap any other Expr of a oneof type that included some Error<T> variant. This worked and seemed very clean as it approximated the same behavior found in languages like Rust, however it came w/ an implementation drawback that would have led to some significant readability problems or worse semantic problems in the language.

Essentially the implementation of the `?` operator was simply inserting a stmts before the current stmt that was creating a new synthetic variable and checking whether it was an error and returning if so. The actual site that used the `?` in source was actually just referencing the synthetic variable created for this purpose. This is fine in some/most cases, but it is problematic in situations where multiple expressions are being evaluated left-to-right and the `?` happens to not be on the far left. It essentially led to a situation where your code would run "out of order" relative to your expectations, as the Expr getting error-checked would run before the Exprs that would appear to need to run first in the source code ordering. Even worse, it would invalidate things like the expected short-circuiting nature of `and`.

As a result of the above problems, I had a few choices largely breaking down between limiting the usage of the `?` operator, or going through a significantly more complex reimplementation to "lift" any Exprs preceding the error-checked one so that they'd all still run before the error-checked Expr.... I decided that the drawbacks of limiting the `?` operator usage are actually not significant enough to warrant the complexity of the fully complete reimplementation. In fact I've been able to largely convince myself that this more limited approach is actually a bit more in the spirit of "Claro" as the place where this early return happens is now appreciably more obvious, whereas the previous `?` approach is somewhat easier to overlook. I'm sure that this will be one syntactical design decision that future users may dislike, but I'm willing to accept that.

So instead of:
```
var foo = mightFail(...)?;
```
you'll do:
```
var foo ?= mightFail(...);
```
which works around the ordering issues I described earlier as this type of thing is no longer possible:
```
if (sideEffectingFn(...) && mightFail(...)?) {
  ...
}
```