Builtin NothingType
This is a very simple addition that simply provides a builtin type to represent "the absence of any significant data". The type is called `NothingType` simply to avoid confusing the type with the literal value `nothing`. This is only builtin to avoid having users redefine this type all over the place.

It's useful for type definitions like below:
```
newtype Maybe<T> : oneof<T, NothingType>
```