For-Loops!
Kinda unbelievable that Claro hasn't had this until now as this was extremely simple to implement, but here we are! Claro will never support C-style for-loops, instead we have "enhanced for-loops".

For now, there's support for iterating over lists/sets/maps however, in the future Claro will allow iteration over *any* type for which some `Iterator` related contract(s) is/are impl'd.