List/Set/Map Comprehension Exprs!!
This is huge for Claro's usability and I'm really thankful for Python inspiring this elegant feature.

Still some more work to come back to in order to support nested comprehensions.

Side Note: The grammar has continued growing and so I needed to make yet another small reduction to the `UPPERLIMIT` of productions per generated function that JCUP uses so that I can still workaround Java's code size limitation on the generated parser.