Quick-Fix for Comprehension
Needed to ensure that even identifier refs at the outermost comprehension are handled in the $NestedComprehensionState to avoid any non-effectively-final lambda capture restrictions.

Made sure to handle this in such a way that if the only identifier refs are in the outermost comprehension's collection expr, then the $NestedComprehensionState isn't generated to avoid the unecessary overhead.