Graph Procedures are Guaranteed Data Race Free!!!
This culminates my planned work on Claro's *Fearless Concurrency* story!

This was shockingly simple to actually implement now that Claro models data mutability at the type system level, but this is now the culmination of over a year of planning!! Claro Graph Procedures are now **GUARANTEED** to have no data races as all underlying graph threads are only capable of using *deeply-immutable* data as input/output. This restriction means that simply by construction you can have a certain guarantee that it is impossible to run into an in-memory data race in Claro!! (Of course you may still have races on data that Claro can't track e.g. in files or across a network, but that's inherently unavoidable).

Now, and forever, the official response I will have for anyone who says something like "But I know what I'm doing, just let me use mutable data however I want!" will be the following: "No you don't. And the rest of us that have the humility to acknowledge that we all make mistakes sometimes and would like to operate in a paradigm where we can assume the absence of this entire class of errors will keep working in Claro - you really should be using Rust or Go if you have some sufficiently advanced, low-level, or convoluted concurrency needs, and you already should've known that."

In the future, Claro may allow the use of mutable data wrapped within AtomicReference<T> or something like that (this would be necessary to enable things like in-memory RPC caches). Once Claro goes in that direction it may technically still be possible to use the atomics incorrectly as I likely won't be able to design around that, but at the very least for now there are no possible races, and even in the future it would only be possible with data explicitly marked "dangerous" via the AtomicReference<T> wrapping. We'll see, no hard plans there yet.