Mutability Coercion on `copy()`!
You are able to seamlessly convert to/from mutable/immutable variants when using copy(). This is a unique capability that you're granted out-of-the-box by means of Claro's copy() having language-level support. This is an extremely useful convenience feature that enables easily working within Claro's deep-immutability restrictions for Graph Procedures that would otherwise cause significant inconvenience. It may of course be advisable to rather design your implementation so as to avoid the need for copying at all, but this gives you options.