Statically Reject Possible Data Races via Transitive Injection of Mutable Data
This is one of those deep cuts that programmers would easily run into on accident as these data races are over faraway data that you otherwise can't see by simply looking at the graph procedure code in isolation. This is unfortunately hard to create a very clear error message for, but at least it gets caught. In the future the error message should be improved so that users are much more easily able to track the path to fixing the problem.