Update Builtin Json Parsing to Support Maps
Now, you're able to parse to maps instead of just to structs as the prior limitation to structs wasn't a good fit for handling JSON as it actually is in use on the web. All kinds of APIs are regularly sending JSON object keys that are either dynamically determined, or contain aribtrary symbols that are not valid Claro identifiers. In either case, it'd be ridiculous to simply reject these JSON formats.