Claro Supports Basic Http Servers!!!!
This is actually a monumental achievement towards my long time overarching goal to make Claro an excellent option for easily building web services that are nearly *guaranteed* to scale well! Claro's builtin support for HttpServers is an incredibly powerful foundation as it's now the only language in the world that can provide static compile-time guarantees that your service is non-blocking and making full use of the multiple cores available on your machine!

LOTS more to come here as Claro continues to blur the line between language and platform.