Initial Support for Match Statements Over Primitives and Oneofs!
This is the first glimmering of the last significant control flow construct currently planned for Claro. Match constructs currently are more or less just a convenient switch statement. The one feature that takes this beyond switch, is simply the fact that you can use this over arbitrary oneofs to exhaustively match and narrow the actual concrete type of the oneof at runtime. This feature alone already hands down makes error values much cleaner to work with as you can get some statically validated confidence that all error cases are handled somehow.

Coming in the future (though it will be a significant challenge so will probably take a good bit of time) will be the ability to push this feature to the extreme by enabling structural pattern matching over nested data! Can't wait for that. For me, that will solidify Claro as an extremely ergonomic language to work with.