Pattern Matching Now Supports Non-Exhaustiveness Warnings!
This is hands down one of the coolest features in the entire language! Massive thanks to this paper "Warnings for pattern matching" (by Luc Maranget) hosted at http://moscova.inria.fr/~maranget/papers/warn/warn007.html. This is the first time I've actually read an academic paper and *ACTUALLY* made use of the information for something Claro related. What a world! Super grateful for that work to be accessible and *ACTUALLY* comprehensible unlike most PL academic papers I've (tried to) read.

I'm only implementing a subset of the work in the paper as I am not currently supporting or-patterns, and I haven't figured out how to use the paper's recommendations on "Uselessness" checks, so likely I'll be coming back to this. But exhaustiveness is by far the most essential element as this will serve Claro's Error-handling story crucially.

Made good progress towards simplifying the implementation to all follow the same general process of flattening/validating patterns and codegen, but much more work to do on pattern matching in terms of significant implementation simplification needed.