Pattern Matching Now Supports Wildcard Bindings for Type Literal Patterns!
This is a big help as it prevents you from needing to manually destructure the value all over again w/in the case body.

Made good progress towards simplifying the implementation to all follow the same general process of flattening/validating patterns and codegen, but much more work to do on pattern matching in terms of significant implementation simplification needed.