Pattern Matching Now Permits the Presence of Unmatchable Values
This exists solely so that you're not disqualified from matching against some nested value solely because some portion of it doesn't support matching. Now, you can simply define matching patterns over the parts of the matched value that actually support it, while wildcards are required to be used in the place of all unsupported types. This is super nice to have because from this point forward, match stmts may be missing some features, but they are at least usable and shouldn't break on any input types.

Note that this includes support for matching "Opaque" User-defined types which are not allowed to be implicitly unwrapped. This check preserves the semantic opaqueness of these particular types, even though technically the compiler would be able to generate the code to do the matching.

Made good progress towards simplifying the implementation to all follow the same general process of flattening/validating patterns and codegen, but much more work to do on pattern matching in terms of significant implementation simplification needed.