Claro Now Supports "Atom" Types!
Atoms are an interesting idea I noticed in Portcullis by Jake Wood from the CoRecursive Slack group (although this is an idea that's implemented in various more mainstream languages such as Elixir). The concept is to create a type that has only a SINGLE value, itself. For example, if the idea of "the absence of data" were expressed as an atom, then it'd be cleaner than the current separation between the type `NothingType` and the value `nothing`.

E.g.
```
atom Nothing

newtype Tree<T> : struct {
  val: T,
  left: oneof<Tree<T>, Nothing>,   # Using `Nothing` as a type.
  right: oneof<Tree<T>, Nothing>
}

var myTree =
  Tree({
    val = 10,
    left =
      Tree({
        val = 20,
        left = Nothing,            # Using `Nothing` as a value.
        right = Nothing
      }),
    right = Nothing
  });
```

The interesting thing about these atoms is that they're "self-describing" in some sense. You could imagine that they represent some more meaningful value, but considering they're the only one of their type in the universe, it's somewhat pointless to *actually* carry that data around, after all, that data really is just some *interpretation* of the presence of such a type.

This is interesting in its own right beyond just the trivial example of making a better representation for NothingType/nothing, as this can also be the foundation of something like an "enum" that you would find in Java/C/etc.

E.g.
```
atom OutOfMemory
atom TimeLimitExceeded
atom PermissionDenied
atom UnknownInternalError

newtype MyErrorsEnum : oneof<
  OutOfMemory,
  TimeLimitExceeded,
  PermissionDenied,
  UnknownInternalError
>
```

Sidenote: As atoms are going to be particularly useful things to match on as oneof-variants. This CL includes some work at ensuring there's some initial level of support for atoms within match-case patterns. As a result I ended up finding some minor MatchStmt bugs and applying some fixes within this CL.