Support Multi-File Programs! (NO MODULE SYSTEM YET)
Amazing that it's taken this long time finally add support for this basic organizational functionality, but here it is! Note that THIS IS NOT ADDING SUPPORT FOR A SYSTEM JUST YET. For now, there's no namespacing or scoping introduced by using multiple files. This is actually by design - for namespacing you should utilize Claro's upcoming module system once ready. The intention here is to give a very lightweight mechanism for organization, setting a very low bar for splitting things into multiple files for readability's sake.

Conceptually, from a design perspective I very well could have implemented a namespacing scheme for accessing cross-file definitions, but I actually see the lack of this namespacing as a heavy motivator *FOR* using Claro's module system once it's introduced. This is an interesting choice, creating something slightly bad intentionally for the sake of pushing people towards the better solution even though that better solution takes some work to make use of. Let's see how the strategy plays out. (Even once full the blown module system is in place, I still think having this lightweight multi-file support available is a huge win for the sake of Claro scripting)

This is a really exciting change as this will not only allow me to write somewhat larger Claro programs in a better organized way, but it will more importantly also enable me to start implementing a "stdlib" using straight up Claro defined in some hidden file! This is going to be a massive improvement over the current approach to adding builtin procedures to the grammar and creating an AST node for it.