First Inkling of Implementing Claro's StdLib in Native Claro!
This builds off of the previous commit that enabled multi-file Claro programs by defining a new `builtin_types.claro` file that exports the few builtin type definitions that are represented as newtypes or atoms. This is a *massive* quality of life improvement for me over needing to manually construct types in some setup stage, as it's very verbose to define types explicitly in Java source. It's now as simple as using Claro source code and leveraging all of the prior work on Claro to parse these stdlib builtin types and put them in scope!

Another really nice thing this enabled was me deleting NothingType from being hardcoded in BaseType.java and Types.java. I was also able to delete the NothingTerm.java synthetic AST Node. The icing on top was deleting multiple grammar productions in the parser to correspond with this.

As I come up with more stdlib functionality, I'll continue to include the functionality in files like these.