Support PrivilegedInlineJavaStmt For Use By The Stdlib Itself
While it won't be very hard for people to figure out ways around this, the intention is for this unsafe functionality to be used by the internal Claro stdlib itself and not by anyone else. In the future if Java interop is ever offered, it will *not* be as unchecked as this.

This CL comes along the reimplementation of `consumer append<T>(l: [T], t: T);` as a stdlib procedure instead of as a compiler intrinsic. This allowed for a significant amount of code deletion which I'm extremely pleased with! This does however add some slight overhead that didn't previously exist as previously with the compiler intrinsic, this operation was inlined at the callsite. This leads me to believe that in the future Claro should likely have some mechanism for specifying that a procedure be "inlined" to avoid the unnecessary procedure call indirection. I don't care about this for now.

Note that any hopes of Riju getting an updated version of Claro are out the window unless I want to jump through some hoops to rewrite Claro's setup integration.