First Pass At CLI For Compiling/Running Claro Programs w/o Bazel
This CL implements two new scripts claroc.sh and claro.sh that compile and run claro programs from the command line respectfully. The major workhorse here is claroc.sh which is doing some major heavy lifting running the claro compiler executable Jar over the given .claro file(s) and then packaging up the outputted .class files into a fat jar alongside all of the necessary java libs that are needed by the codegen'd java program. The fantastic thing about this script (in my eyes) is that it produces a *single executable Jar file* that can simply be carried around and run (ideally) anywhere that Java's installed. After you're done compiling, you actually already have a completely portable executable artifact on your hands!

This is currently in a very experimental state that in the very literal sense "works on my machine" and won't work on anyone else's w/o modifications to the CLI scripts. But this is a huge start!