Update the claroc.sh CLI Script to Accept Paths to Necessary Jars Via Env Vars
This change makes it so that the script doesn't just work on my machine anymore. Now, you should be able to "simply" configure the paths to `claro_compiler_binary_deploy.jar` and `claro_builtin_java_deps_deploy.jar` as ENV vars `CLARO_COMPILER_PATH` and `CLARO_BUILTIN_DEPS_PATH` respectfully. For now, you get the first Jar by simply grabbing it from the latest Github release on clarolang.com, but the deps Jar still needs to be setup for releases in the Github release action. That's still a TODO.

The nice thing here is that until I actually manage to find a more legit approach to "installing" these CLIs, once I get these CLIs and the necessary Jars included as part of the Github Actions release process, you should be able to simply download the release zip, unzip and then start compiling from that directory w/o even specifying any env vars as the script will auto-check for the jars in the curr dir.

This current script should be ready for use w/in the Riju configuration to get Riju updates back to a working state.