Update the claroc.sh CLI Script to Assume Installation Structure
This change makes it so that the script can be used as a standalone shell command, e.g.:
`$ claroc *.claro`
rather than the more annoying
`$ ./claroc.sh *.claro`

However, this requires some additional setup steps that weren't previously required. I think that this is acceptable setup for now considering the steps can still be copy-pasted and an installation script can be provided later on.

Setup (Assuming /usr/local/bin is in your PATH):
1. mkdir /usr/local/lib/claro
2. curl -L https://github.com/JasonSteving99/claro-lang/releases/latest/download/claro-cli-install.tar.gz > /usr/local/lib/claro/claro-cli-install.tar.gz
3. tar xzf /usr/local/lib/claro/claro-cli-install.tar.gz -C /usr/local/lib/claro
4. rm /usr/local/lib/claro/claro-cli-install.tar.gz
5. ln -s /usr/local/lib/claro/claroc.sh /usr/local/bin/claroc
6. ln -s /usr/local/lib/claro/claro.sh /usr/local/bin/claro

This current script should be ready for use w/in the Riju configuration to get Riju updates back to a working state.