Begin Work on Claro's First-Class Module System!
This is a very simple CL that's just laying the groundwork to enable parsing very trivial procedure-signature-only module api files. This includes a simple test java program that reads a module api file off stdin and then validates that it's able to actually able to parse the procedure signatures out of the module api file.

This is the fundamental unit of indirection that will allow Claro to type check that a unit of Claro code (a Module) exports/binds the set of procedure signatures that it claims to. Once I have this validation, Claro will be able to generate a non-executable Jar that can be utilized as a dependency of another Claro Module.. there's a ton more groundwork to go and I'm not going to explain the entire design here. Get ahold of my notebook for the details ;P.