Continue Work on Claro's First-Class Module System!
Now Claro has some level of support for signaling that you wish to compile a set of .claro source files against a .claro_module_api file to compile a validated module. Claro will parse the signatures present in the module api file, and validate that the given sources actually do provide implementations of the exported bindings. For now, once this is validated, Claro just emits a codegen'd .java file as it always has. This is not the final form as this would not be able to handle exported generic procedures as they still need to be monomorphized according to actual usage by dependents.

I wanted to continue straight towards producing a .claro_module file format utilizing protobuf, but it appears I need to finally sort out upgrading Bazel to the latest version (I'm super far out of date) first. So I'll next up figure out how to start making use of Bazelisk.