Begin Work on Removing Lombok Dep From Project
This is a massive annoyance as I never should've had a lombok dep in the first place. It was worth experimenting with but didn't turn out to be particularly useful to me. I'm particularly removing this now because after the migration to the latest Bazel version I suddenly need to use the --nojava_header_compilation flag on all of my bazel build/run commands simply because Lombok has some obnoxious failure mode w/o it. Lombok's not fully removed in this CL but this required a ton of manual refactoring so I'm checking this in now.