Update JavaSourceCompilerBackend.java to Use Real Command Line Parser
This not only cleans up the implementation, but is particularly useful to ensure that the compiler's parsing CLI args in the format that Bazel will be passing them for the new in-development claro_binary() rule impl to replace the claro_binary() macro.

This change impacted the args accepted, so I made some changes to the claroc.sh script as well to appropriately handle this.