Re-Implement Claro's claro_binary() Macro as a First-Class Bazel Rule!
This seemingly insignificant step is one more step towards clearing away the remaining obstacles in the way that are preventing me from working directly on Claro's First-Class Module System. But this is a great step! I learned a ton about making Bazel Rules (which is still not saying a lot, this was much harder than I wish it was, and I still have a lot to learn regarding propagating transitive depsets).

After this, I'll be able to continue working on the claro_library rule which is now completely freed up from any legitimate dependencies, so I can go straight into working on converting it into a new claro_moduel() rule that will generate .claro_module outputs, and will propagate depsets to be used by the claro_binary() rule.