Re-Implement Claro's claro_library() Macro as claro_module() in terms of _invoke_claro_compiler() Bazel Rule
The claro_library() macro was already completely freed up from any legitimate dependencies, so I re-implemented it as claro_module() that will generate .claro_module outputs, and will soon be used to propagate depsets to be used by the claro_binary() for "linking". This is the fundamental unit of incrementality in Claro's Bazel integration.