Claro's First-Class Module System is Fully* Functioning!
(*For non-Generic Procedures)

This is a massive moment! This is the first POC that shows the long-planned direct integration with Bazel is in fact a viable long term path!

Now, Claro supports incremental compilation out of the box. And since the only mechanism for creating a new namespace in Claro is by making a new Module using the Bazel rule claro_module(), users are almost literally given no other option than using Bazel "correctly" by creating many small targets to optimize for incrementality.

This task was surprisingly easy, but the much harder task is still to come, in supporting calls to generic procedures defined in other "compilation units" (modules).