Prepare the Demo BuggyBuggies Bot HttpServer to be the Test Case for Claro's First-Class Module System
This server is currently the most significant Claro program in existence, and is the perfect subject of study for actually putting Claro's in development Module system to use. For now, I have destructured the old http_server.claro file into some "realistic" folder structure so that this resembles some sort of "professional project organization". For now, these are all still just pulled in as claro_binary() srcs so everything's still getting dumped into a single compilation unit and a single global namespace. As I continue to flesh out the Module System support, I will progressively migrate each sub-component into a proper claro_module() target and start introducing proper namespacing and increasing the incrementality of this program's build.

This is a great test subject because when all is said and done, the resulting program will be composed of a proper DAG of modules (not just a tree), and there will be modules exporting types/aliases/intitializers/unwrappers AND generic functions that need monomorphizing by upstream callers. So this will occupy me for quite some time as I'm able to progressively flesh out the module system.

One interesting thing that I've done here just as a demonstration of the power of Claro's integration with Bazel is the use of genrules to simply demonstrate programmatically generating .claro sources in a safe and reproducible way. The specific example is a demonstration of simulating "Macros" w/o needing to actually introduce macros into the Claro language.