Update the Demo BuggyBuggies Bot HttpServer to Build a Couple More Modules
This is a great start towards leveraging Claro's First-Class Module system in a "real" program. This is just a start that introduces a few Modules to introduce some namespacing and incrementality to this program's build. The Module created from endpoint_handlers.claro_module_api is currently still hacked together around some duplicated srcs deps on a bunch of .claro files that should definitely be consumed as modules when it's possible. Currently this is prevented because Claro Modules still do not allow the exporting of newtype defs and aliases. This is such a great start though!