Module Api Def Files Now Support Exporting Newtype Def Initializers/Unwrappers
Initializers and Unwrappers blocks are very useful for controlling how your dependents can interact with the types that you expose. The great thing about this separation between .claro and .claro_module_api files is that you're empowered to omit the initializers/unwrappers blocks within the impl so that internally your module implementation is unrestricted in how it may interact with its own types, while dependents must follow stricter rules of your choosing.

A MASSIVE change to namespacing comes along in this same change. Now all NewTypeDefStmt's automatically produce types that contain the unique module descriptor within its type definition at both compile and runtime. This allows two modules to define/export types with the same names but have Claro correctly identify them as distinct.

The one hangup here is that the builtin types `ParsedJson` and `Error` are currently implicitly included in each compilation unit, meaning that some pervasive hacking was necessary to ensure that the parsers (and runtime utilities that manually initialize instances of these types) put every definition of these types into the same (empty) namespace. This is problematic in lots of ways, but will be addressed in the future by factoring all builtin types into a set of real stdlib modules (which is long awaited anyways).